requirements.txt
flask==2.3.3
flask-cors==4.0.0
firebase-admin==6.2.0
# openai==0.28.1  # Removed - using Claude Haiku only
anthropic==0.64.0
requests==2.31.0
redis==4.6.0
python-dotenv==1.0.0
gunicorn==21.2.0
